+++ lib/pages/home/pages/community/widgets/community_card/widgets/community_details/community_details.dart
@@ -1,6 +1,7 @@
 import 'package:Okuna/models/community.dart';
 import 'package:Okuna/pages/home/pages/community/widgets/community_card/widgets/community_details/widgets/community_favorite.dart';
 import 'package:Okuna/pages/home/pages/community/widgets/community_card/widgets/community_details/widgets/community_members_count.dart';
+import 'package:Okuna/pages/home/pages/community/widgets/community_card/widgets/community_details/widgets/community_posts_count.dart';
 import 'package:Okuna/pages/home/pages/community/widgets/community_card/widgets/community_details/widgets/community_type.dart';
 import 'package:flutter/material.dart';
 
@@ -23,11 +24,11 @@ class OBCommunityDetails extends StatelessWidget {
             Flexible(
               child: SizedBox(
                 child: Wrap(
-                  spacing: 10.0,
-                  runSpacing: 10.0,
+                  spacing: 0.0,
                   children: <Widget>[
                     OBCommunityType(community),
                     OBCommunityMembersCount(community),
+                    OBCommunityPostsCount(community),
                     OBCommunityFavorite(community)
                   ],
                 ),