+++ lib/widgets/posts_stream/posts_stream.dart
@@ -435,6 +435,7 @@ class OBPostsStreamState extends State<OBPostsStream>
     setState(() {
       _posts.remove(deletedPost);
       if (_posts.isEmpty) _setStatus(OBPostsStreamStatus.empty);
+      if (deletedPost.isCommunityPost()) deletedPost.community.decrementPostsCount();
     });
   }