+++ CHANGELOG.md
@@ -12,9 +12,12 @@ The change log for the Okuna app.
  - Show suggested communities on account creation
  - Add choose username page
  - Trending posts is now an infinite stream
+ - Add Okuna age to profile
+ - Add posts count to community page
 
 #### Bugs
 - Verify token on invite opened
+- Remove username from forgot password
 
 #### Happy testing 🎉!